1，处理提示语（从数据库配置） 2，增加退出登录接口 3，处理短信验证码逻辑

+++ minigod-cloud-persist/src/main/java/com/minigod/persist/account/mapper/CustomOpenInfoMapper.java
@@ -57,5 +57,7 @@ public interface CustomOpenInfoMapper {
 
     CustomOpenInfo selectOneByPhone(@Param("phone") String phone);
 
-    List<CustomOpenInfo> selectByCaStatus(@Param("caStatus") Byte caStatus);
+    List<CustomOpenInfo> selectByCaStatus(@Param("caStatus")Integer caStatus);
+
+
 }
